Move bootstrap to 'torus machines bootstrap', Gatekeeper is no longer co-process
