--- cmd/spawn_gatekeeper_windows.go
@@ -1,13 +1,0 @@
-package cmd
-
-import (
-	"os/exec"
-	"syscall"
-)
-
-func gatekeeperCommand(executable string) *exec.Cmd {
-	cmd := exec.Command(executable, "gatekeeper", "start", "--foreground", "--daemonize")
-	cmd.SysProcAttr = &syscall.SysProcAttr{}
-
-	return cmd
-}
