+++ gatekeeper/gatekeeper.go
@@ -4,9 +4,7 @@ package gatekeeper
 
 import (
 	"fmt"
-	"os"
 
-	"github.com/nightlyone/lockfile"
 	"github.com/urfave/cli"
 
 	"github.com/manifoldco/torus-cli/api"
@@ -17,22 +15,11 @@ import (
 // Gatekeeper provides the Listener interface for a daemonized HTTP co-process that will act
 // to register machines
 type Gatekeeper struct {
-	lock lockfile.Lockfile
-	g    *http.Gatekeeper
+	g *http.Gatekeeper
 }
 
 // New returns a new Gatekeeper
-func New(ctx *cli.Context, cfg *config.Config, groupShared bool) (g *Gatekeeper, err error) {
-	lock, err := lockfile.New(cfg.GatekeeperPidPath)
-	if err != nil {
-		return nil, fmt.Errorf("Failed to create lockfile object: %s", err)
-	}
-
-	err = lock.TryLock()
-	if err != nil {
-		return nil, fmt.Errorf("Failed to create lockfile[%s]: %s", cfg.GatekeeperPidPath, err)
-	}
-
+func New(ctx *cli.Context, cfg *config.Config) (g *Gatekeeper, err error) {
 	// recovery and cleanup
 	defer func() {
 		if r := recover(); r != nil {
@@ -42,18 +29,11 @@ func New(ctx *cli.Context, cfg *config.Config, groupShared bool) (g *Gatekeeper,
 		}
 	}()
 
-	if groupShared {
-		if err := os.Chmod(string(lock), 0600); err != nil {
-			return nil, err
-		}
-	}
-
 	api := api.NewClient(cfg)
 	http := http.NewGatekeeper(ctx, cfg, api)
 
 	gatekeeper := &Gatekeeper{
-		lock: lock,
-		g:    http,
+		g: http,
 	}
 
 	return gatekeeper, nil
@@ -72,10 +52,6 @@ func (g *Gatekeeper) Run() error {
 
 // Shutdown gracefully stops the HTTP service
 func (g *Gatekeeper) Shutdown() error {
-	if err := g.lock.Unlock(); err != nil {
-		return fmt.Errorf("Could not unlock: %s", err)
-	}
-
 	if err := g.g.Close(); err != nil {
 		return fmt.Errorf("Could not stop gatekeeper: %s", err)
 	}
